--- src/main/java/com/plog/domain/hashtag/service/HashTagServiceImpl.java
@@ -1,68 +1,0 @@
-package com.plog.domain.hashtag.service;
-
-import com.plog.domain.hashtag.entity.HashTag;
-import com.plog.domain.hashtag.entity.PostHashTag;
-import com.plog.domain.hashtag.repository.HashTagRepository;
-import com.plog.domain.hashtag.repository.PostHashTagRepository;
-import com.plog.domain.post.entity.Post;
-import com.plog.domain.post.repository.PostRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class HashTagServiceImpl implements HashTagService {
-
-    private final HashTagRepository hashTagRepository;
-    private final PostHashTagRepository postHashTagRepository;
-    private final PostRepository postRepository;
-
-    @Override
-    @Transactional
-    public void createPostHashTag(Long postId, List<String> tagNames) {
-        Post post = postRepository.getReferenceById(postId);
-        applyTags(post, tagNames); // 공통 로직 호출
-    }
-
-    @Override
-    @Transactional
-    public void updatePostHashTag(Long postId, List<String> tagNames) {
-        postHashTagRepository.deleteAllByPostId(postId);
-
-        Post post = postRepository.getReferenceById(postId);
-        applyTags(post, tagNames); // 공통 로직 호출
-    }
-
-    /**
-     * [공통 로직] 태그 이름을 분석하여 PostHashTag를 생성하고 저장합니다.
-     */
-    private void applyTags(Post post, List<String> tagNames) {
-        if (tagNames == null || tagNames.isEmpty()) return;
-
-        for (String rawName : tagNames) {
-            String normalizedName = normalizeTag(rawName);
-
-            HashTag hashTag = hashTagRepository.findByName(normalizedName)
-                    .orElseGet(() -> hashTagRepository.save(new HashTag(normalizedName)));
-
-            if (!postHashTagRepository.existsByPostIdAndHashTagId(post.getId(), hashTag.getId())) {
-                PostHashTag postHashTag = PostHashTag.builder()
-                        .post(post)
-                        .hashTag(hashTag)
-                        .displayName(rawName)
-                        .build();
-
-                postHashTagRepository.save(postHashTag);
-            }
-        }
-    }
-
-
-
-    private String normalizeTag(String name) {
-        return name.trim().toLowerCase().replace(" ", "_");
-    }
-}
