+++ src/main/java/com/plog/domain/post/controller/PostController.java
@@ -1,7 +1,5 @@
 package com.plog.domain.post.controller;
 
-import com.plog.domain.comment.constant.CommentConstants;
-import com.plog.domain.hashtag.service.HashTagService;
 import com.plog.domain.post.dto.PostCreateReq;
 import com.plog.domain.post.dto.PostInfoRes;
 import com.plog.domain.post.dto.PostListRes;
@@ -51,7 +49,6 @@ import java.net.URI;
 public class PostController {
 
     private final PostService postService;
-    private final HashTagService hashTagService;
 
     /**
      * 새로운 게시물을 생성합니다.
@@ -65,7 +62,6 @@ public class PostController {
             @Valid @RequestBody PostCreateReq request
     ) {
         Long postId = postService.createPost(user.getId(), request);
-        hashTagService.createPostHashTag(postId, request.hashtags());
 
         return ResponseEntity.created(URI.create("/api/posts/" + postId)).build();
     }
