+++ src/test/java/com/plog/domain/post/service/PostServiceTest.java
@@ -2,7 +2,6 @@ package com.plog.domain.post.service;
 
 import com.plog.domain.comment.repository.CommentRepository;
 import com.plog.domain.hashtag.repository.PostHashTagRepository;
-import com.plog.domain.hashtag.service.HashTagService;
 import com.plog.domain.member.entity.Member;
 import com.plog.domain.member.repository.MemberRepository;
 import com.plog.domain.post.dto.PostCreateReq;
@@ -51,9 +50,6 @@ public class PostServiceTest {
 
     @Mock
     private PostHashTagRepository postHashTagRepository;
-
-    @Mock
-    private HashTagService hashTagService;
 
     @Test
     @DisplayName("게시글 저장 시 마크다운이 제거된 요약글이 자동 생성")
